chore: rename PreloadIssuesAlert -> StoreErrorsAlert; update unit test;
